feat(linux): collapse Settings into four pages
Merge the thin Adwaita panes so later destinations can grow by group without a six-page window.
