+++ platforms/linux/settings-gtk/src/settings_window/general.rs
@@ -1,14 +1,15 @@
-//! "Chung" page: general preferences (launch at login) + config backup.
+//! "Chung" page: session preference, config backup, and About.
 
 use adw::prelude::*;
 use adw::{PreferencesGroup, PreferencesPage, PreferencesWindow, SwitchRow};
 
 use crate::settings::Settings;
 
+mod about;
 mod transfer;
 
 pub(super) fn page(window: &PreferencesWindow) -> PreferencesPage {
-    let s = Settings::load();
+    let settings = Settings::load();
     let page = PreferencesPage::builder()
         .title("Chung")
         .icon_name("preferences-system-symbolic")
@@ -20,15 +21,19 @@ pub(super) fn page(window: &PreferencesWindow) -> PreferencesPage {
     let row = SwitchRow::builder()
         .title("Khởi động cùng phiên đăng nhập")
         .subtitle("Bộ gõ do desktop quản lý tự khởi động; tuỳ chọn này chỉ được lưu lại.")
-        .active(s.launch_at_login)
+        .active(settings.launch_at_login)
         .build();
     row.connect_active_notify(|row| {
         let on = row.is_active();
-        Settings::update(|s| s.launch_at_login = on);
+        Settings::update(|settings| settings.launch_at_login = on);
     });
     group.add(&row);
     page.add(&group);
 
     page.add(&transfer::group(window));
+
+    let about_group = PreferencesGroup::new();
+    about_group.add(&about::row(window));
+    page.add(&about_group);
     page
 }
