+++ platforms/linux/settings-gtk/src/settings_window/typing.rs
@@ -1,0 +1,23 @@
+//! "Cách gõ" page: how keys become Vietnamese and how the word reaches the app.
+//! Each group lives in `typing/` so a later option is one file plus one `page.add`.
+
+mod delivery;
+mod method;
+mod smart;
+
+use adw::prelude::*;
+use adw::PreferencesPage;
+
+use crate::settings::Settings;
+
+pub(super) fn page() -> PreferencesPage {
+    let settings = Settings::load();
+    let page = PreferencesPage::builder()
+        .title("Cách gõ")
+        .icon_name("input-keyboard-symbolic")
+        .build();
+    page.add(&method::group(&settings));
+    page.add(&smart::group(&settings));
+    page.add(&delivery::group(&settings));
+    page
+}
